Add stable-ts to requirements.txt and update video.json structure
Split main.py into multiple files
