+++ requirements.txt
@@ -10,4 +10,5 @@ torchaudio==2.0.1+cu117
 torchvision==0.15.1+cu117
 mkdocs-material
 openai-whisper
+stable-ts
 tiktok-uploader
