+++ video.json
@@ -1,19 +1,11 @@
-[
-    {
-        "series": "Crazy facts that you did not know",
-        "part": "Part 4",
-        "outro": "Follow us for more",
-        "text": "Did you know that there are more possible iterations of a game of chess than there are atoms in the observable universe? The number of possible legal moves in a game of chess is around 10^120!",
-        "tags": [
-            "chess",
-            "facts",
-            "crazy"
-        ]
-    },
-    {
-        "series": "Crazy facts that you did not know",
-        "part": "Part 5",
-        "outro": "Follow us for more",
-        "text": "Hi, this is a test and it will be a video!"
-    }
-]
+{
+    "series": "Crazy facts that you did not know",
+    "part": "Part 4",
+    "outro": "Follow us for more",
+    "text": "Did you know that there are more possible iterations of a game of chess than there are atoms in the observable universe? The number of possible legal moves in a game of chess is around 10^120!",
+    "tags": [
+        "chess",
+        "facts",
+        "crazy"
+    ]
+}
